seperated rule for GoL/updating cell
