+++ src/org/bitstorm/gameoflife/GameOfLifeGrid.java
@@ -14,47 +14,9 @@ import java.util.Hashtable;
  *
  * @author Edwin Martin
  */
-public class GameOfLifeGrid implements CellGrid {
-	private int cellRows;
-	private int cellCols;
-	private int generations;
-	private static Shape[] shapes;
-	/**
-	 * Contains the current, living shape.
-	 * It's implemented as a hashtable. Tests showed this is 70% faster than Vector.
-	 */
-	private Hashtable currentShape;
-	private Hashtable nextShape;
-	/**
-	 * Every cell on the grid is a Cell object. This object can become quite large.
-	 */
-	private Cell[][] grid;
-
-	/**
-	 * Contructs a GameOfLifeGrid.
-	 * 
-	 * @param cellCols number of columns
-	 * @param cellRows number of rows
-	 */
+public class GameOfLifeGrid extends CellGrid {
 	public GameOfLifeGrid(int cellCols, int cellRows) {
-		this.cellCols = cellCols;
-		this.cellRows = cellRows;
-		currentShape = new Hashtable();
-		nextShape = new Hashtable();
-
-		grid = new Cell[cellCols][cellRows];
-		for ( int c=0; c<cellCols; c++)
-			for ( int r=0; r<cellRows; r++ )
-				grid[c][r] = new Cell( c, r );
-	}
-
-	/**
-	 * Clears grid.
-	 */
-	public synchronized void clear() {
-		generations = 0;
-		currentShape.clear();
-		nextShape.clear();
+		super(cellCols,cellRows);
 	}
 
 	/**
@@ -112,134 +74,5 @@ public class GameOfLifeGrid implements CellGrid {
 				setCell( cell.col, cell.row, true );
 			}
 		}
-	}
-	
-	/**
-	 * Adds a new neighbour to a cell.
-	 * 
-	 * @param col Cell-column
-	 * @param row Cell-row
-	 */
-	public synchronized void addNeighbour(int col, int row) {
-		try {
-			Cell cell = (Cell)nextShape.get( grid[col][row] );
-			if ( cell == null ) {
-				// Cell is not in hashtable, then add it
-				Cell c = grid[col][row];
-				c.neighbour = 1;
-				nextShape.put(c, c);
-			} else {
-				// Else, increments neighbour count
-				cell.neighbour++;
-			}
-		} catch (ArrayIndexOutOfBoundsException e) {
-			// ignore
-		}
-	}
-	
-	/**
-	 * Get enumeration of Cell's
-	 * @see org.bitstorm.gameoflife.CellGrid#getEnum()
-	 */
-	public Enumeration getEnum() {
-		return currentShape.keys();
-	}
-
-	/**
-	 * Get value of cell.
-	 * @param col x-coordinate of cell
-	 * @param row y-coordinate of cell
-	 * @return value of cell
-	 */
-	public synchronized boolean getCell( int col, int row ) {
-		try {
-			return currentShape.containsKey(grid[col][row]);
-		} catch (ArrayIndexOutOfBoundsException e) {
-			// ignore
-		}
-		return false;
-	}
-
-	/**
-	 * Set value of cell.
-	 * @param col x-coordinate of cell
-	 * @param row y-coordinate of cell
-	 * @param c value of cell
-	 */
-	public synchronized void setCell( int col, int row, boolean c ) {
-		try {
-			Cell cell = grid[col][row];
-			if ( c ) {
-				currentShape.put(cell, cell);
-			} else {
-				currentShape.remove(cell);
-			}
-		} catch (ArrayIndexOutOfBoundsException e) {
-			// ignore
-		}
-	}
-	
-	/**
-	 * Get number of generations.
-	 * @return number of generations
-	 */
-	public int getGenerations() {
-		return generations;
-	}
-	
-	/**
-	 * Get dimension of grid.
-	 * @return dimension of grid
-	 */
-	public Dimension getDimension() {
-		return new Dimension( cellCols, cellRows );
-	}
-
-	/**
-	 * Resize grid. Reuse existing cells.
-	 * @see org.bitstorm.gameoflife.CellGrid#resize(int, int)
-	 */
-	public synchronized void resize(int cellColsNew, int cellRowsNew) {
-		if ( cellCols==cellColsNew && cellRows==cellRowsNew )
-			return; // Not really a resize
-
-		// Create a new grid, reusing existing Cell's
-		Cell[][] gridNew = new Cell[cellColsNew][cellRowsNew];
-		for ( int c=0; c<cellColsNew; c++)
-			for ( int r=0; r<cellRowsNew; r++ )
-				if ( c < cellCols && r < cellRows )
-					gridNew[c][r] = grid[c][r];
-				else
-					gridNew[c][r] = new Cell( c, r );
-
-		// Copy existing shape to center of new shape
-		int colOffset = (cellColsNew-cellCols)/2;
-		int rowOffset = (cellRowsNew-cellRows)/2;
-		Cell cell;
-		Enumeration enumi;
-		nextShape.clear();
-		enumi = currentShape.keys();
-		while ( enumi.hasMoreElements() ) {
-			cell = (Cell) enumi.nextElement();
-			int colNew = cell.col + colOffset;
-			int rowNew = cell.row + rowOffset;
-			try {
-				nextShape.put( gridNew[colNew][rowNew], gridNew[colNew][rowNew] );
-			} catch ( ArrayIndexOutOfBoundsException e ) {
-				// ignore
-			}
-		}
-
-		// Copy new grid and hashtable to working grid/hashtable
-		grid = gridNew;
-		currentShape.clear();
-		enumi = nextShape.keys();
-		while ( enumi.hasMoreElements() ) {
-			cell = (Cell) enumi.nextElement();
-			currentShape.put( cell, cell );
-		}
-		
-		cellCols = cellColsNew;
-		cellRows = cellRowsNew;
 	}
 }
